net: mvneta: use page pool API for sw buffer manager
Use the page_pool api for allocations and DMA handling instead of
__dev_alloc_page()/dma_map_page() and free_page()/dma_unmap_page().

The page_pool API offers buffer recycling capabilities for XDP but
allocates one page per packet, unless the driver splits and manages
the allocated page.
Although XDP is not a part of the driver yet, the current implementation
is allocating one page per packet, thus there's no performance penalty from
using the API.

A previous patch introduces buffer recycling using the page_pool API
The driver memcpy for pakets < 256b and it's recycle tricks are not needed
anymore.

This patch also fixes a driver bug where the memory is copied prior to it's
syncing for the cpu (in the < 256b case).

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>
Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>