net: core: add recycle capabilities on skbs via page_pool API
Place a new member 'mem_info' of type struct xdp_mem_info
right after dma_addr_t in struct page that holds all the necessary
information for page recycling. It's the drivers responsibility to fill
in the information.

During skb_release_data() check and recycle the data buffer of the skb and
it's frags

The 'mem_info' name is chosen as this is not strictly tied to XDP,
although the XDP return infrastructure is used.

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>
Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>