bpf/cpumap: make sure frame_size for build_skb is aligned if headroom isn't
The frame_size passed to build_skb must be aligned, else it is
possible that the embedded struct skb_shared_info gets unaligned.

For correctness make sure that xdpf->headroom in included in the
alignment. No upstream drivers can hit this, as all XDP drivers provide
an aligned headroom.  This was discovered when playing with implementing
XDP support for mvneta, as this Marvell ARM64 platform didn't like
doing atomic operations on an unaligned skb_shinfo(skb)->dataref address.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>